added ui for updating and reviewing books

+++ app/Http/Controllers/BookController.php
@@ -36,15 +36,52 @@ class BookController extends Controller
      */
     public function store(Request $request)
     {
+        $max_size = (int)ini_get('upload_max_filesize') * 1000;
 
         //Validate
         $validator = Validator::make($request->all(), [
             'title' => 'required',
-            'description' => 'required|min:8'
+            'description' => 'required|min:8',
+            'cover_img'=>'image|nullable|max:' . $max_size,
+            'book'=>'file|nullable|max:' . $max_size
         ]);
 
         if ($validator->fails()) {
             return response()->json(['error' => 'title and description is required'], 400);
+        }
+
+        // handle file upload
+        if($request->hasFile('cover_img')){
+            // get file with extenstion
+            $fileNamewithExt = $request->file('cover_img')->getClientOriginalName();
+            // get file name only
+            $filename = pathinfo($fileNamewithExt, PATHINFO_FILENAME);
+            // get extension only
+            $extension = $request->file('cover_img')->getClientOriginalExtension();
+            // filename to store
+            $fileNameToStoreImg = $filename.'_'.time().'.'.$extension;
+            // upload image
+            $path = $request->file('cover_img')->storeAs('public/cover_images', $fileNameToStoreImg);
+        }else{
+            // no file was uploaded
+            $fileNameToStoreImg = 'no_cover_img.jpg';
+        }
+
+        // handle file upload
+        if($request->hasFile('book')){
+            // get file with extenstion
+            $fileNamewithExt = $request->file('book')->getClientOriginalName();
+            // get file name only
+            $filename = pathinfo($fileNamewithExt, PATHINFO_FILENAME);
+            // get extension only
+            $extension = $request->file('book')->getClientOriginalExtension();
+            // filename to store
+            $fileNameToStoreBook = $filename.'_'.time().'.'.$extension;
+            // upload image
+            $path = $request->file('book')->storeAs('public/books', $fileNameToStoreBook);
+        }else{
+            // no file was uploaded
+            $fileNameToStoreBook = 'no_book.png';
         }
 
         try {
@@ -53,6 +90,8 @@ class BookController extends Controller
             $book->user_id = $user->id;
             $book->title = $request->input('title');
             $book->description = $request->input('description');
+            $book->cover_img = $fileNameToStoreImg;
+            $book->book = $fileNameToStoreBook;
             $user->books()->save($book);
         }catch(Exception $ex){
             return response()->json(['error' => $ex->getMessage()], 400);
@@ -71,6 +110,12 @@ class BookController extends Controller
     public function show($id)
     {
         $book = Book::find($id);
+        if($book == null){
+            return response()->json(['error' => 'book not found'], 404);
+        }
+
+        $book->cover_img = '/storage/cover_images/'.$book->cover_img;
+        $book->book = '/storage/books/'.$book->book;
         $book->with(['user', 'ratings']);
         return (new BookResource($book))->response()->setStatusCode(200);
     }
@@ -85,23 +130,74 @@ class BookController extends Controller
     public function update(Request $request, $id)
     {
 
+        $book = Book::find($id);
+        if($book == null){
+            return response()->json(['error' => 'book not found'], 404);
+        }
+
+        if ($request->user()->id !== $book->user_id) {
+            return response()->json(['error' => 'You can only edit your own books.'], 403);
+        }
+
+        $max_size = (int)ini_get('upload_max_filesize') * 1000;
+
+        error_log(json_encode($request->input('title')));
+
         //Validate
         $validator = Validator::make($request->all(), [
             'title' => 'required',
-            'description' => 'required|min:8'
+            'description' => 'required|min:8',
+            'cover_img'=>'image|nullable|max:' . $max_size,
+            'book'=>'file|nullable|max:' . $max_size
         ]);
 
         if ($validator->fails()) {
             return response()->json(['error' => 'title and description is required'], 400);
         }
 
-        $book = Book::find($id);
-
-        if ($request->user()->id !== $book->user_id) {
-            return response()->json(['error' => 'You can only edit your own books.'], 403);
+        // handle file upload
+        if($request->hasFile('cover_img')){
+            // get file with extenstion
+            $fileNamewithExt = $request->file('cover_img')->getClientOriginalName();
+            // get file name only
+            $filename = pathinfo($fileNamewithExt, PATHINFO_FILENAME);
+            // get extension only
+            $extension = $request->file('cover_img')->getClientOriginalExtension();
+            // filename to store
+            $fileNameToStoreImg = $filename.'_'.time().'.'.$extension;
+            // upload image
+            $path = $request->file('cover_img')->storeAs('public/cover_images', $fileNameToStoreImg);
+        }else{
+            // no file was uploaded
+            $fileNameToStoreImg = 'no_cover_img.jpg';
         }
 
-        $book->update($request->only(['title', 'description']));
+        // handle file upload
+        if($request->hasFile('book')){
+            // get file with extenstion
+            $fileNamewithExt = $request->file('book')->getClientOriginalName();
+            // get file name only
+            $filename = pathinfo($fileNamewithExt, PATHINFO_FILENAME);
+            // get extension only
+            $extension = $request->file('book')->getClientOriginalExtension();
+            // filename to store
+            $fileNameToStoreBook = $filename.'_'.time().'.'.$extension;
+            // upload image
+            $path = $request->file('book')->storeAs('public/books', $fileNameToStoreBook);
+        }else{
+            // no file was uploaded
+            $fileNameToStoreBook = 'no_book.png';
+        }
+
+        $book->title = $request->input('title');
+        $book->description = $request->input('description');
+        if($request->hasFile('cover_img')){
+            $book->cover_img = $fileNameToStoreImg;
+        }
+        if($request->hasFile('book')){
+            $book->book = $fileNameToStoreBook;
+        }
+        $book->save();
 
         return (new BookResource($book))->response()->setStatusCode(200);
     }
@@ -115,6 +211,9 @@ class BookController extends Controller
     public function destroy($id)
     {
         $book = Book::find($id);
+        if($book == null){
+            return response()->json(['error' => 'book not found'], 404);
+        }
         if ($book->delete()) {
             return (new BookResource($book))->response()->setStatusCode(200);
         }
